+++ lib/td_dd/data_structures/structure_notes/policy.ex
@@ -11,17 +11,17 @@ defmodule TdDd.DataStructures.StructureNotes.Policy do
 
   def authorize(action, claims, %DataStructure{domain_ids: domain_ids} = ds)
       when action in [
+             :ai_suggestions,
              :create,
              :delete,
              :deprecate,
              :edit,
-             :publish,
+             :history,
              :publish_draft,
+             :publish,
              :reject,
              :submit,
-             :unreject,
-             :history,
-             :ai_suggestions
+             :unreject
            ] do
     Bodyguard.permit?(DataStructures, :view_data_structure, claims, ds) and
       Permissions.authorized?(claims, permission(action), domain_ids)
@@ -29,15 +29,15 @@ defmodule TdDd.DataStructures.StructureNotes.Policy do
 
   def authorize(_action, _, _params), do: false
 
+  defp permission(:ai_suggestions), do: :ai_structure_note
   defp permission(:create), do: :create_structure_note
   defp permission(:delete), do: :delete_structure_note
   defp permission(:deprecate), do: :deprecate_structure_note
   defp permission(:edit), do: :edit_structure_note
   defp permission(:history), do: :view_structure_note_history
-  defp permission(:publish), do: :publish_structure_note
   defp permission(:publish_draft), do: :publish_structure_note_from_draft
+  defp permission(:publish), do: :publish_structure_note
   defp permission(:reject), do: :reject_structure_note
   defp permission(:submit), do: :send_structure_note_to_approval
   defp permission(:unreject), do: :unreject_structure_note
-  defp permission(:ai_suggestions), do: :ai_structure_note
 end