+++ lib/td_dq_web/views/rule_view.ex
@@ -46,6 +46,7 @@ defmodule TdDqWeb.RuleView do
     |> Map.take([
       :active,
       :business_concept_id,
+      :business_concept_name,
       :deleted_at,
       :description,
       :domain_id,