+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [TD-6938] Add Grant Requests Search scroll and inserted_at filter
 - [TD-6780] Add Tag and Structure Tag Seach controller and context
+- [TD-6982] Added SSL and ApiKey configuration for Elasticsearch
 
 ## [6.15.0] 2024-11-27